+++ Sources/Plasma/PubUtilLib/plDrawable/plDynaRippleVSMgr.h
@@ -59,7 +59,7 @@ protected:
 
     bool        IRippleFromShape(const plPrintShape* shape, bool force=false) override;
 
-    int         INewDecal() override;
+    size_t      INewDecal() override;
 
     virtual bool        ICheckRTMat();
